Add Makefile.uk
This is a port of axtls to Unikraft as an external library. It requires
libc and lwip to work.

To run ssltest.c as a unit test, you'll need pthread support, e.g.,
add the following line to the LIBS variable in the Makefile:

   ...$(UK_LIBS)/lwip:$(UK_LIBS)/pthread-embedded:
         $(UK_LIBS)/newlib:$(UK_LIBS)/axtls...

Note that you'll need a filesystem in place and the certificate files for
all tests to pass.

Signed-off-by: Felipe Huici <felipe.huici@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>